Refactor preinst and apt-get/pip install
All base-image related apt-get commands (including pre-install config,
as well as the package install) have been put together in the same RUN
instruction.

The same has been done with the LAVA-related apt-get commands on this
patch. They can't be merged because the requirements are using a
specific key that needs to be downloaded with curl (that's installed in
the previous step. Additionally, LAVA apt-get commands have been moved
up in the Dockerfile, allowing that RUN instruction to be invalidated
fewer times, because it does not depend on the files that have been
copied from the host.

We have also moved apt sources to an external .list file, that's copied
from the host to the container during build.

Signed-off-by: Guilherme Campos Camargo <guicc@profusion.mobi>