Cleanup proxy
Prior to this patch, http_proxy was being set and saved during build
time according to the http_proxy values of the host.

This approach has at least two shortcomings, that are explained below:

1. Any other user that wanted to use a prebuilt image in their host
would be unable to do so, because the proxy configuration is matching
the host of the builder and not the host of the user.

2. A user of the image would be able to read the proxy configuration
that had been set during build from `/etc/default/jenkins` or the
apt-get config file. The proxy config may easily contain sensitive
information as the proxy user/password that were set on the host that
built the image.

In order to avoid the problems pointed out above, we chose to:

1. Remove the ARG/ENV instantiations of http_proxy from the Dockerfile
and use only the builtin docker ARG http_proxy. This will prevent the
proxy variable to be available in a running container and in the image
history.

2. Instead of using the apt-get global config file, we're just passing
http_proxy as a local environment variable to the RUN instruction in
which apt-get is being executed.

3. Jenkins /etc/default/jenkins JAVA_ARGS (that contain the proxy
information) is now being updated by a script that's been source by
/etc/default/jenkins itself. This means that, whenever `service jenkins`
is called, that script will run and append the proxy configurations to
the JAVA_ARGS at runtime, not replacing the JAVA_ARGS that are stored in
the file.

4. http_proxy variables are being passed as arguments to `docker create`
as well. By doing that, we make sure that the entrypoint will be running
with the http_proxy values of the user.

Signed-off-by: Guilherme Campos Camargo <guicc@profusion.mobi>

Conflicts:
	frontend-install/setup/jenkins/setup.sh
	fuego-host-scripts/docker-build-image.sh