Refactor plugin installation script
On this patch we're trying to remove the explicit sleeps from the plugin
installation script, that are necessary for waiting for Jenkins
start/restart and the installation of the plugins.

With this change we hope that faster hosts (or hosts with faster
networks) will be able to finish Jenkins plugins installation sooner.

We also took the chance to remove the call to sudo, that are not
necessary for using jenkins-cli.

Signed-off-by: Guilherme Campos Camargo <guicc@profusion.mobi>