check and modify docker jenkins uid and gid on startup
By default, set the docker jenkins uid and gid to the
uid and gid of the user on the host who is creating
the docker image (write it into the Fuego global config
file).  When the image starts, use entrypoint.sh to
reset the uid and gid, and chmod pertinent directories,
if the uid has changed.

Signed-off-by: Tim Bird <tim.bird@sony.com>