Install Fuego Release Test dependencies and add fuego-test board
A new session has been added to the Dockerfile, mainly responsible for
installing the Docker binary, pexpect and seleniumHQ (that are needed
for Fuego to test a new Fuego release).

Also, a fuego-test board (a simple loopback that points to localhost)
has been added.

The create scripts have also been modified to allow mapping the hosts'
dockerd socket into the container. By doing this we allow Fuego and the
host to share the same docker daemon.