Allow Fuego to run without the need of mounting fuego-ro/core/rw
This change is necessary for allowing Fuego to be encapsulated into a
Docker image, not requiring any other resource and scripts for running
it.

The user may choose to run fuego in standalone mode or in persistency
mode (as explained in README_DOCKERHUB).

By running in standalone mode (without any volume or mountpoint passed
to Fuego, to store data), fuego-ro/rw is copied from the template stored
in this repo, that contains a few boards, documentation and the expected
file structure. In this mode fuego-core is cloned from the official
fuego-core master branch.

By running in persistency mode, Fuego will try to use the provided
fuego-ro/fuego-rw and fuego-core, re-generating them only if they're not
present into the mountpoint/volume.

We're also adding the possibility to pass the expected jenkins uid/gid
as a docker run argument in order to allow jenkins to read/write from
any directory that the user may want to bind mount as fuego-home

Signed-off-by: Guilherme Campos Camargo <guicc@profusion.mobi>