Remove and re-create /var/cache/jenkins to avoid permission problems
On some hosts /var/cache/jenkins may become not accessible by the
jenkins user after the uid of the user gets mapped to the host's uid,
even after properly calling chown on it.

We did not discover why this happens, but by removing and re-creating
the directory, the permission issue is solved.

Signed-off-by: Guilherme Campos Camargo <guicc@profusion.mobi>