Use gosu instead of su on entrypoint.sh to avoid unexpected exits
Gosu is often used in the Docker official images as a way to execute a
command as a user different than the root without explicitely setting
USER on Dockerfile. This is useful, for example, for entrypoints.sh in
which several commands need to be run as root while the main command
does not need to be, as in our case.

Su/sudo could also be used, but in the official best practices, their
use is discouraged given to their "unpredictable TTY and
signal-forwarding behavior, that can cause problems".

By replacing `su` with `gosu`, the problem of unexpected container exits
seems to have been solved.

Signed-off-by: Guilherme Campos Camargo <guicc@profusion.mobi>